split network interface into two smaller classes: NetworkResolver and NetworkEventQueue, wrapping fetch requests and broadcast/poll respectively. handler and dereference function calls now pass keyword arguments now, ctx and rid/kobj
